// types/job/review.ts
export interface Review {
  id: string;
  interviewId: string;
  reviewerId: string;
  rating?: number | undefined;
  comment?: string | undefined;
  createdAt: string;
}

/**
 * ---------------------
 * CRUD Responses
 * ---------------------
 */
export interface ReviewResponse {
  success: boolean;
  message?: string | undefined;
}

export interface ListReviewsResponse {
  reviews: Review[];
  total: number;
  page: number;
  limit: number;
}

export interface DeleteReviewResponse {
  success: boolean;
  message: string;
}

export interface AllReviewsWithMetaResponse {
  data: Review[];
  meta: PaginationMeta | undefined;
}

export interface SingleReviewResponse {
  message?: string | undefined;
  review: Review | undefined;
}

/**
 * ---------------------
 * CRUD Requests
 * ---------------------
 */
export interface ListReviewsRequest {
  page: number;
  limit: number;
  rating?: number | undefined;
  reviewerId?: string | undefined;
  interviewId?: string | undefined;
}

export interface GetReviewRequest {
  reviewId: string;
}

export interface UpdateReviewRequest {
  reviewId: string;
  rating?: number | undefined;
  comment?: string | undefined;
}

export interface DeleteReviewRequest {
  reviewId: string;
}

export interface GetAllReviewsRequest {
  basicQuery: BasicQuery | undefined;
}

export interface CreateReviewRequest {
  interviewId: string;
  reviewerId: string;
  rating?: number | undefined;
  comment?: string | undefined;
}

export const REVIEWS_PACKAGE_NAME = "reviews";

/**
 * ---------------------
 * CRUD Service
 * ---------------------
 */

export interface ReviewServiceClient {
  createReview(request: CreateReviewRequest): Observable<ReviewResponse>;

  getReview(request: GetReviewRequest): Observable<SingleReviewResponse>;

  getAllReviews(request: GetAllReviewsRequest): Observable<AllReviewsWithMetaResponse>;

  updateReview(request: UpdateReviewRequest): Observable<ReviewResponse>;

  deleteReview(request: DeleteReviewRequest): Observable<DeleteReviewResponse>;

  listReviews(request: ListReviewsRequest): Observable<ListReviewsResponse>;
}